# console/public/openapi.yaml
openapi: "3.1.0"
info:
  title: SecureClaim AI — Showcase API
  version: "0.1.0"
  description: |
    Read-only public API surface for the SecureClaim AI Resilience Console.
    Exposes the attack-defense matrix, defense-pattern catalog, live architecture
    metadata, session replays, and the playground submission endpoint.

    All endpoints are public (no authentication required) except where noted.
    Playground submission (`POST /showcase/playground`) is rate-limited to
    10 requests per IP per minute.

    ## Transport types

    | Type | Prefix | Notes |
    |------|--------|-------|
    | REST | `/showcase/` | Standard JSON; ISR-cached where noted |
    | SSE  | `/sse/`       | `text/event-stream`; no caching |
    | WS   | `/ws/`        | See [WebSocket channels](#section/WebSocket-channels) below |

    ## WebSocket channels

    `ws://{host}/ws/playground`

    Bidirectional stream for the Attack Playground.  The client sends a
    `PlaygroundSubmission` JSON frame and the server pushes back
    `PlaygroundEvent` frames (one per defense layer evaluated) until a
    `verdict` event closes the stream.

    **Client → server frame:**
    ```json
    { "type": "submit", "payload": "<attack text>", "templateId": "string|null" }
    ```

    **Server → client frames:**
    ```json
    { "type": "layer_result", "layer": "P1", "outcome": "blocked|partial|pass",
      "detail": "string", "ts": "ISO-8601" }
...
      "traceId": "uuid",     "ts": "ISO-8601" }
...
    ```

...
servers:
...
paths:
...
  /showcase/matrix:
    get:
      operationId: getMatrix
      tags: [matrix]
      summary: Full attack-defense matrix
      description: |
        Returns all 79 attack categories with aggregate test counts, the
        defense patterns that address each, and a current status indicator.
        Response is ISR-cached; cache is revalidated on every CI webhook.
      parameters:
        - $ref: "#/components/parameters/MatrixClassFilter"
        - $ref: "#/components/parameters/PatternFilter"
      responses:
        "200":
          description: Matrix data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatrixResponse"
...
components:
...
  parameters:
...
    MatrixClassFilter:
      name: class
      in: query
      description: Filter rows to a specific test-result class
      schema:
        $ref: "#/components/schemas/MatrixClass"
    PatternFilter:
      name: pattern
      in: query
      description: Filter to attacks addressed by a specific pattern ID
      schema:
        $ref: "#/components/schemas/PatternId"

  # ─── Responses ───────────────────────────────────────────────────────────────
...
  schemas:
...
    MatrixClass:
      type: string
      enum: [BLOCKED, PARTIAL, SUCCESSFUL, UNTESTED]
      description: |
        Aggregate test outcome for an attack category.
        `BLOCKED` — all recorded attempts blocked; `PARTIAL` — blocked in
        most but not all scenarios; `SUCCESSFUL` — at least one unblocked
        attempt recorded; `UNTESTED` — no test data yet.
    PatternId:
      type: string
      pattern: "^P(1[0-2]|[1-9])$"
      description: Defense pattern identifier P1–P12
      example: P8
...
    MatrixRow:
      type: object
      required: [attackId, name, class, patterns, blockedCount, partialCount, successfulCount]
      properties:
        attackId:
          type: integer
          minimum: 1
          maximum: 79
        name:
          type: string
          description: Short attack category name (from the 79-category taxonomy)
          example: "Prompt Injection via PDF"
        class:
          $ref: "#/components/schemas/MatrixClass"
        patterns:
          type: array
          items:
            $ref: "#/components/schemas/PatternId"
          description: Defense patterns that address this attack
        blockedCount:
          type: integer
          minimum: 0
        partialCount:
          type: integer
          minimum: 0
        successfulCount:
          type: integer
          minimum: 0
        lastTestedAt:
          type: string
          format: date-time
          nullable: true
    MatrixResponse:
      type: object
      required: [rows, generatedAt]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/MatrixRow"
        generatedAt:
          type: string
          format: date-time
...